+++ blocks/magazine-articles/magazine-articles.css
@@ -155,6 +155,7 @@
   height: 100%;
   width: 100%;
   object-fit: cover;
+  aspect-ratio: 16 / 9;
 }
 
 .block.magazine-articles article picture.default-image img {
