+++ scripts/services/magazine.service.js
@@ -1,4 +1,4 @@
-import { getLocale } from '../common.js';
+import { getLocale, getOrigin, getTextLabel } from '../common.js';
 import { fetchSearchData, topicSearchQuery, magazineSearchQuery, TENANT } from '../search-api.js';
 
 /**
@@ -220,4 +220,40 @@ export const fetchRecommendedArticles = async ({
     console.error('Error fetching recommended articles:', error);
     return null;
   }
+};
+
+const defaultAuthor = getTextLabel('defaultAuthor');
+const defaultReadTime = getTextLabel('defaultReadTime');
+/**
+ * Extract the classes of a block and in case there is a 'limit-X' set, extract it as a number
+ * @param {Object} item - The raw article item that comes from opensearch
+ * @returns {Object} - The parsed object without the metadata field and some other properties
+ */
+export const parseArticleData = (item) => {
+  // Add a check for item and item.metadata in case 'item' is falsy
+  if (!item || !item.metadata) {
+    console.warn('parseArticleData received invalid item or metadata:', item);
+    return;
+  }
+
+  const isImageLink = (link) => `${link}`.split('?')[0].match(/\.(jpeg|jpg|gif|png|svg|bmp|webp)$/) !== null;
+
+  const getDefaultImage = () => {
+    const logoImageURL = '/media/logo/media_10a115d2f3d50f3a22ecd2075307b4f4dcaedb366.jpeg';
+    return getOrigin() + logoImageURL;
+  };
+
+  const { article, image } = item.metadata;
+  const filterTag = ['category', 'topic', 'truck'].map((key) => item.metadata.article[key]).filter(Boolean);
+
+  return {
+    ...item.metadata,
+    filterTag,
+    author: article.author || defaultAuthor,
+    image: isImageLink(image) ? getOrigin() + image : getDefaultImage(),
+    path: item.metadata?.url,
+    readingTime: /\d+/.test(article.readTime) ? article.readTime : defaultReadTime,
+    isDefaultImage: !isImageLink(image),
+    category: article.category,
+  };
 };
